% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.smle}
\alias{coef.smle}
\alias{coef.selection}
\title{Extract coefficients from fitted model}
\usage{
\method{coef}{smle}(object, ...)

\method{coef}{selection}(object, ...)
}
\arguments{
\item{object}{Returned object from either the function \code{\link{SMLE}} or \code{\link{smle_select}}.}

\item{...}{This argument is not used and listed for method consistency.}
}
\value{
Fitted coefficients based on the screened or selected model specified in the object.
}
\description{
Extract coefficients from fitted model for either a \code{'smle'} or \code{'selection'} object.
}
\examples{

set.seed(1)
Data<-Gen_Data(n=100, p=5000, family = "gaussian", correlation="ID")
fit<-SMLE(Y=Data$Y, X=Data$X, k=9, family = "gaussian")
coef(fit)
fits<-smle_select(fit)
coef(fits)
}
