% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.smle.R
\name{logLik.smle}
\alias{logLik.smle}
\title{Extract log-likelihood}
\usage{
\method{logLik}{smle}(object, ...)
}
\arguments{
\item{object}{Object of class \code{'smle'} or \code{'sdata'}.}

\item{...}{Forwarded arguments.}
}
\value{
Value of log-likelihood
}
\description{
Extract log-likelihood
}
