% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.smle.R
\name{predict.smle}
\alias{predict.smle}
\title{Prediction based on SMLE screening and selection}
\usage{
\method{predict}{smle}(object, newdata = NULL, type = c("link", "response"), ...)
}
\arguments{
\item{object}{A fitted object of class \code{'smle'} as the output from
SMLE.}

\item{type}{The type of prediction required. Type "link" leads to a prediction 
on the scale of linear predictors (i.e. linear combination of retained features); 
type \code{'response'} leads to a prediction on the mean value of the response.
When \code{'family=gaussian'} in \code{'smle'}, the two predication types are equivalent.}

\item{...}{Further arguments pass to \code{predict.glm()}.}

\item{newx}{Matrix of new values for x at which predictions are to be made,
without the intercept term. If omitted, the function returns the fitted 
response values based on the training data and the retained features in \code{'smle'}.}
}
\value{
The object returned depends on type.
}
\description{
This function makes prediction for the response in new data based on the 
features retained in \code{'smle'} object.
}
\examples{

set.seed(123.456)

Data_sim<-Gen_Data(n= 200, p =1000, correlation="AR",family = "gaussian")

fit<-SMLE(Data_sim$Y,Data_sim$X, family = "gaussian" , k=10)

predict(fit , type ="link")

}
