% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sdata.R
\name{print.sdata}
\alias{print.sdata}
\title{Print a sdata object}
\usage{
\method{print}{sdata}(x, ...)
}
\arguments{
\item{x}{Fitted '\code{sdata}' object.}

\item{...}{This argument is not used and listed for method consistency.}
}
\value{
No return value.
}
\description{
This functions prints the basic information of a data set generated by
Gen_data.
}
