% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.selection.R
\name{print.selection}
\alias{print.selection}
\title{Print a selection object}
\usage{
\method{print}{selection}(x, ...)
}
\arguments{
\item{x}{Fitted '\code{selection}' object.}

\item{...}{This argument is not used and listed for method consistency.}
}
\value{
No return value.
}
\description{
This functions prints the retained subset of a SMLE object.
}
