% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals.R
\docType{data}
\name{pvals}
\alias{pvals}
\title{p values of synthetic genetic association study data set}
\format{
An object of class \code{data.frame} with 10031 rows and 4 columns.
}
\usage{
data(pvals)
}
\description{
The first column is the chromosome number. The second columns is SNP name. 
 The third column is the genomic position of the SNP on the whole data set.
 The marginal p-values of each SNPs is pre-calculated and saved in the fourth column.
}
\keyword{datasets}
