% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.selection.R
\name{summary.selection}
\alias{summary.selection}
\title{Summary a selection object from smle_select}
\usage{
\method{summary}{selection}(object, ...)
}
\arguments{
\item{object}{Fitted \code{'selection'} object.}

\item{...}{This argument is not used and listed for method consistency.}
}
\value{
No return value, called for side effects.
}
\description{
This function prints a summary of a \code{'selection'} object.
In particular, it gives the selected features along with their re-fitted model coefficients.
For reference, it also shows the values of the selection criterion used in selection for all candidate models.
}
\examples{
Data<-Gen_Data(correlation="MA",family = "gaussian")
fit<-SMLE(Data$Y,Data$X,k=20,family = "gaussian")
E<-smle_select(fit)
summary(E)

}
