% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.smle}
\alias{coef.smle}
\alias{coef.selection}
\title{Extract coefficient from fitted model}
\usage{
\method{coef}{smle}(object, ...)

\method{coef}{selection}(object, ...)
}
\arguments{
\item{object}{Fitted model}

\item{...}{Forwarded arguments}
}
\value{
Coefficients after screening or selection.
}
\description{
Extract coefficient from fitted model
}
