% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.smle}
\alias{logLik.smle}
\alias{logLik.selection}
\title{Extract log-likelihood}
\usage{
\method{logLik}{smle}(object, ...)

\method{logLik}{selection}(object, ...)
}
\arguments{
\item{object}{Object of class \code{'smle'} or \code{'sdata'}.}

\item{...}{Forwarded arguments.}
}
\value{
Returns an object of class logLik. This is a number with at least one attribute, "df" (degrees of freedom), giving the number of (estimated) parameters in the model.
}
\description{
Extract log-likelihood
}
