% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.smle}
\alias{print.smle}
\alias{print.selection}
\alias{print.summary.smle}
\alias{print.sdata}
\title{Print a SMLE object}
\usage{
\method{print}{smle}(x, ...)

\method{print}{selection}(x, ...)

\method{print}{summary.smle}(x, ...)

\method{print}{sdata}(x, ...)
}
\arguments{
\item{x}{Fitted '\code{smle}' object.}

\item{...}{This argument is not used and listed for method consistency.}
}
\value{
No return value.
}
\description{
This functions prints the retained subset of a SMLE object.
}
