% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sdata.R
\name{print.sdata}
\alias{print.sdata}
\title{Print function for simulated data}
\usage{
\method{print}{sdata}(x, ...)
}
\arguments{
\item{x}{\code{"sdata"} object from Gen_Data function.}

\item{...}{Other parameter for print function.}
}
\value{
No return value, called for side effects.
}
\description{
This functions prints a summary of a data set sumulated with Gen_data.
In particular, it prints which features were selected to be causal, their coefficients and about the correlation structure.
}
\examples{
Data<-Gen_Data(family ="binomial",correlation = "ID")
cor(Data$X[,1:10])
print(Data)
}
