% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.selection.R
\name{print.selection}
\alias{print.selection}
\title{Print a selection object from smle_select}
\usage{
\method{print}{selection}(x, ...)
}
\arguments{
\item{x}{Fitted \code{"selection"} object.}

\item{...}{Other parameter to print.}
}
\value{
No return value, called for side effects.
}
\description{
This function prints a summary of a selection object.
In particular, it gives the selected features along with their re-fitted model coefficient.
For reference, it also shows the values of the selection criterion used in selection for all candidate models.
}
\examples{
Data<-Gen_Data(correlation="MA",family = "gaussian")
fit<-SMLE(Data$Y,Data$X,k=20,family = "gaussian")
E<-smle_select(fit)
print(E)

}
