% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.smle.R
\name{print.smle}
\alias{print.smle}
\title{Print a SMLE object from SMLE}
\usage{
\method{print}{smle}(x, ...)
}
\arguments{
\item{x}{Fitted \code{"smle"} object.}

\item{...}{Other parameter for print.}
}
\value{
No return value, called for side effects.
}
\description{
This functions prints a summary of a SMLE object.
In particular, it shows the features retained after SMLE-screening
and the related convergence information.
}
\examples{
Data<-Gen_Data(correlation="MA",family = "gaussian")
fit<-SMLE(Data$Y,Data$X,k=20,family = "gaussian")
print(fit)
}
